move Pixbuf entriely to gdkpixbuf, add methods GetWidth and GetHeight
