+++ gdk/gdk.go
@@ -453,78 +453,6 @@ func (v *Window) Native() *C.GdkWindow {
 }
 
 /*
- * GdkPixbuf
- */
-
-// Pixbuf is a representation of GDK's GdkPixbuf.
-type Pixbuf struct {
-	*glib.Object
-}
-
-func wrapPixbuf(obj *glib.Object) *Pixbuf {
-	return &Pixbuf{obj}
-}
-
-// Native() returns a pointer to the underlying GdkWindow.
-func (v *Pixbuf) Native() *C.GdkPixbuf {
-	if v == nil || v.GObject == nil {
-		return nil
-	}
-	p := unsafe.Pointer(v.GObject)
-	return C.toGdkPixbuf(p)
-}
-
-// PixbufNewFromFile() is a wrapper around gdk_pixbuf_new_from_file().
-func PixbufNewFromFile(fileName string) (*Pixbuf, error) {
-	var cerr *C.GError
-	cstr := C.CString(fileName)
-	defer C.free(unsafe.Pointer(cstr))
-	c := C.gdk_pixbuf_new_from_file(cstr, &cerr)
-	if c == nil {
-		defer C.g_error_free(cerr)
-		return nil, errors.New(C.GoString((*C.char)(C.error_get_message(cerr))))
-	}
-	obj := &glib.Object{glib.ToGObject(unsafe.Pointer(c))}
-	p := wrapPixbuf(obj)
-	runtime.SetFinalizer(obj, (*glib.Object).Unref)
-	return p, nil
-}
-
-// PixbufNewFromFileAtSize() is a wrapper around gdk_pixbuf_new_from_file_at_size().
-func PixbufNewFromFileAtSize(fileName string, width, height int) (*Pixbuf, error) {
-	var cerr *C.GError
-	cstr := C.CString(fileName)
-	defer C.free(unsafe.Pointer(cstr))
-	c := C.gdk_pixbuf_new_from_file_at_size(cstr, C.int(width), C.int(height), &cerr)
-	if c == nil {
-		defer C.g_error_free(cerr)
-		return nil, errors.New(C.GoString((*C.char)(C.error_get_message(cerr))))
-	}
-	obj := &glib.Object{glib.ToGObject(unsafe.Pointer(c))}
-	p := wrapPixbuf(obj)
-	runtime.SetFinalizer(obj, (*glib.Object).Unref)
-	return p, nil
-}
-
-// PixbufNewFromFileAtScale() is a wrapper around gdk_pixbuf_new_from_file_at_scale().
-func PixbufNewFromFileAtScale(fileName string, width, height int, preserveAspectRatio bool) (*Pixbuf, error) {
-	var cerr *C.GError
-	cstr := C.CString(fileName)
-	defer C.free(unsafe.Pointer(cstr))
-	c := C.gdk_pixbuf_new_from_file_at_scale(cstr, C.int(width), C.int(height), gbool(preserveAspectRatio), &cerr)
-	if c == nil {
-		defer C.g_error_free(cerr)
-		return nil, errors.New(C.GoString((*C.char)(C.error_get_message(cerr))))
-	}
-	obj := &glib.Object{glib.ToGObject(unsafe.Pointer(c))}
-	p := wrapPixbuf(obj)
-	runtime.SetFinalizer(obj, (*glib.Object).Unref)
-	return p, nil
-}
-
-// TODO gdk_pixbuf_get_file_info, resource and stream related functions.
-
-/*
  * GdkRectangle
  */
